Move the creation of Runtime dot graph to a separate submodule
