+++ oak/server/rust/oak_runtime/src/runtime/introspect.rs
@@ -14,7 +14,7 @@
 // limitations under the License.
 //
 
-use crate::runtime::Runtime;
+use crate::runtime::{graph::DotGraph, Runtime};
 use hyper::{
     service::{make_service_fn, service_fn},
     Body, Error, Method, Request, Response, Server, StatusCode,
